commit pre code

+++ Template/Project/DataModels/ProductModel.cs
@@ -7,13 +7,13 @@ public class ProductModel
     public string Description { get; set; }
 
     public string Keywords { get; set; }
-    public decimal Price { get; set; }
+    public double Price { get; set; }
 
 
     public ProductModel(){ }
 
 
-    public ProductModel(string productname, string description, string keywords, decimal price)
+    public ProductModel(string productname, string description, string keywords, double price)
     {
         ProductName = productname;
         Description = description;
